--- decho/src/App.css
@@ -1,40 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.column{
-  /* max-width:500px; */
-  margin: 0 auto;
-}
-
-.landing{
-  background: linear-gradient(#349DF3, #ffffff);
-  width: 100vw;
-  /* padding: 0 50vmax; */
-  /* border:0px; */
-}
-
-body{
-  min-height: 100vh;
-  /* background: linear-gradient(45deg, #349DF3 0 50%, #151D2B 50% 100%); */
-}
-
-.logo{
-  min-width: 50vmin;
-  max-width: 100vw;
-  min-height: 50vmin;
-  max-height: 100vh;
-  background: linear-gradient(45deg, #349DF3 0 50%, #151D2B 50% 100%);
-  padding: 0 50vmax;
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
