+++ decho/src/Components/App.css
@@ -1,0 +1,19 @@
+.App {
+  text-align: center;
+}
+
+body {
+  min-height: 100vh;
+  /* background: linear-gradient(45deg, #349DF3 0 50%, #151D2B 50% 100%); */
+}
+
+.App-header {
+  background-color: #282c34;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+  color: white;
+}
