+++ decho/src/Components/App.js
@@ -1,15 +1,12 @@
 import React, { Component } from 'react';
-
-import 'Components/App';
-import LandingPage from 'Pages/LandingPage';
-import TimelinePage from 'Pages/TimelinePage';
+import LandingPage from 'Pages/Landing/LandingPage';
+import TimelinePage from 'Pages/Timeline/TimelinePage';
 import { getOneOppFromServer } from 'Utils/getOpp';
-import '../App.css';
+import 'Components/App.css';
 // import { getOneOppFromServer } from '../Utils/getOpp';
 // import LandingPage from '../Pages/Landing//LandingPage';
 // import TimelinePage from '../Pages/Timeline/TimelinePage';
 // import { getOneOppFromServer } from '../Utils/getOpp';
-
 
 class App extends Component {
   state = {
@@ -17,26 +14,26 @@ class App extends Component {
     currentLean: 'right',
     opp: '',
   };
-  
-  changePage = (page) => {
-   return (currentLean) =>{
 
-    getOneOppFromServer(currentLean)
-      .then((opp) => {
+  changePage = (page) => {
+    return (currentLean) => {
+      getOneOppFromServer(currentLean).then((opp) => {
         this.setState({
           currentPage: page,
-          currentLean: currentLean || this.state.currentLean ,
-          opp
-        })
-      })
-   }
-  }
+          currentLean: currentLean || this.state.currentLean,
+          opp,
+        });
+      });
+    };
+  };
   render() {
     const { currentPage } = this.state;
     return (
       <div className="App">
-        {currentPage === 'landing' && <LandingPage changePage={ this.changePage('timeline') } />}
-        {currentPage === 'timeline' && <TimelinePage politicalLean={this.state.currentLean} partisan={this.state.opp}/>}
+        {currentPage === 'landing' && <LandingPage changePage={this.changePage('timeline')} />}
+        {currentPage === 'timeline' && (
+          <TimelinePage politicalLean={this.state.currentLean} partisan={this.state.opp} />
+        )}
       </div>
     );
   }
