--- decho/src/Components/ShowProfile.js
@@ -1,20 +1,0 @@
-import React, { Component } from 'react'
-import { Timeline } from 'react-twitter-widgets'
-
-export default class ShowProfile extends Component {
-  render() {
-    return (
-      <div>
-        <Timeline
-          dataSource={{
-            sourceType: 'profile',
-            screenName: 'BarackObama'
-          }}
-          options={{
-            height: '800'
-          }}
-        />
-      </div>
-    )
-  }
-}
